saving code for js and views still a work in progress testing is next when completed
