+++ public/js/logout.js
@@ -7,9 +7,8 @@ const logout = async () => {
     if (response.ok) {
       document.location.replace('/');
     } else {
-      alert(response.statusText);
+      alert('Logout failed.');
     }
   };
   
-  document.querySelector('#logout').addEventListener('click', logout);
-  
+document.querySelector('#logout').addEventListener('click', logout);
