+++ public/js/signup.js
@@ -1,4 +1,22 @@
-async function signupFormHandler(event) {
+const signupFormHandler = async (event) => {
     event.preventDefault();
-    const userName = document.querySelector('#username-signup').value.trim();
-}
+
+    const username = document.querySelector('#username-signup').value.trim();
+    const password = document.querySelector('#password-signup').value.trim();
+
+    if (username && password) {
+        const response = await fetch('/api/users', {
+            method: 'POST',
+            body: JSON.stringify({ username, password }),
+            headers: { 'Content-Type' : 'application/json' },
+        });
+
+        if (response.ok) {
+            document.location.replace('/dashboard');
+        } else {
+            alert('Sign-up failed.');
+        }
+    }
+};
+
+document.querySelector('.signup-form').addEventListener('submit', signupFormHandler);
