+++ server.js
@@ -1,16 +1,18 @@
 const express = require('express');
 const sequelize = require('./config/connection');
-const SequelizeStore = require('connect-session-sequelize')(session.Store);
-const session = require('express-session');
-const exphbs = require('express-handlebars');
 const routes = require('./controllers');
+const path = require('path');
+
+const session = require('express-session');
+
 const helpers = require('./utils/helpers');
+const exphbs = require('express-handlebars');
+const hbs = exphbs.create({ helpers });
 
 const app = express();
 const PORT = process.env.PORT || 3001;
 
-// const hbs = exphbs.create({ helpers });
-
+const SequelizeStore = require('connect-session-sequelize')(session.Store);
 const sess = {
     secret: 'Super secret secret',
     cookie: {
