--- views/home.handlebars
@@ -1,5 +1,0 @@
-<article class="">
-    {{#each - as |-| }}
-    {{> - -}}
-    {{/each}}
-</article>
