+++ views/homepage.handlebars
@@ -1,0 +1,5 @@
+<article class="post">
+    {{#each posts as |post| }}
+    {{> post-info posts}}
+    {{/each}}
+</article>
