+++ views/partials/comments.handlebars
@@ -1,0 +1,12 @@
+<div class="comments">
+    {{ #each this }}
+    <section class="comment">
+        <div class="meta">
+            {{ user.username }} on {{ format_date created_at }}
+        </div>
+        <div class="text">
+            {{ comment_contents }}
+        </div>
+    </section>
+    {{ /each }}
+</div>
